fix(android): do not set properties on recycled proxy (#11302)
* refactor accessibility
* update MenuItemProxy with new changes

Fixes TIMOB-27496
