+++ android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIDialog.java
@@ -285,7 +285,7 @@ public class TiUIDialog extends TiUIView
 						}
 						ViewCompat.setImportantForAccessibility(listView, importance);
 					} else {
-						listView.setContentDescription(getProxy().composeContentDescription());
+						listView.setContentDescription(composeContentDescription());
 					}
 				}
 			}
@@ -338,7 +338,7 @@ public class TiUIDialog extends TiUIView
 			if (listView != null) {
 				int importance = ViewCompat.IMPORTANT_FOR_ACCESSIBILITY_AUTO;
 				if (proxy != null) {
-					listView.setContentDescription(proxy.composeContentDescription());
+					listView.setContentDescription(composeContentDescription());
 					Object propertyValue = proxy.getProperty(TiC.PROPERTY_ACCESSIBILITY_HIDDEN);
 					if (propertyValue != null && TiConvert.toBoolean(propertyValue)) {
 						importance = ViewCompat.IMPORTANT_FOR_ACCESSIBILITY_NO;
