+++ android/titanium/src/java/org/appcelerator/kroll/KrollProxy.java
@@ -32,7 +32,6 @@ import android.os.Handler;
 import android.os.Looper;
 import android.os.Message;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.util.Pair;
 
 import org.json.JSONObject;
@@ -974,57 +973,6 @@ public class KrollProxy implements Handler.Callback, KrollProxySupport, OnLifecy
 		}
 
 		return krollObject.fireEvent(source, event, krollData, bubbles, reportSuccess, code, message);
-	}
-
-	/**
-	 * Our view proxy supports three properties to match iOS regarding
-	 * the text that is read aloud (or otherwise communicated) by the
-	 * assistive technology: accessibilityLabel, accessibilityHint
-	 * and accessibilityValue.
-	 *
-	 * We combine these to create the single Android property contentDescription.
-	 * (e.g., View.setContentDescription(...));
-	 */
-	public String composeContentDescription()
-	{
-		if (properties == null) {
-			return null;
-		}
-
-		final String punctuationPattern = "^.*\\p{Punct}\\s*$";
-		StringBuilder buffer = new StringBuilder();
-		String label = TiConvert.toString(properties.get(TiC.PROPERTY_ACCESSIBILITY_LABEL));
-		String hint = TiConvert.toString(properties.get(TiC.PROPERTY_ACCESSIBILITY_HINT));
-		String value = TiConvert.toString(properties.get(TiC.PROPERTY_ACCESSIBILITY_VALUE));
-
-		if (!TextUtils.isEmpty(label)) {
-			buffer.append(label);
-			if (!label.matches(punctuationPattern)) {
-				buffer.append(".");
-			}
-		}
-
-		if (!TextUtils.isEmpty(value)) {
-			if (buffer.length() > 0) {
-				buffer.append(" ");
-			}
-			buffer.append(value);
-			if (!value.matches(punctuationPattern)) {
-				buffer.append(".");
-			}
-		}
-
-		if (!TextUtils.isEmpty(hint)) {
-			if (buffer.length() > 0) {
-				buffer.append(" ");
-			}
-			buffer.append(hint);
-			if (!hint.matches(punctuationPattern)) {
-				buffer.append(".");
-			}
-		}
-
-		return buffer.toString();
 	}
 
 	public void firePropertyChanged(String name, Object oldValue, Object newValue)
